+++ async-progress.js
@@ -1,1 +1,1 @@
-export const ASYNC_PROGRESS_PREFIX = "__MCP_ASYNC_PROGRESS__ ";
+export * from "./src/runtime/async-progress.js";
