+++ test/unit/validation.test.mjs
@@ -21,8 +21,8 @@ describe("package.json files allowlist", () => {
   test("every local JS module imported by src/index.js is in the files allowlist", () => {
     const entrypointSrc = fs.readFileSync(path.join(root, "src", "index.js"), "utf8");
     const localImports = [
-      ...entrypointSrc.matchAll(/^\s*import\b(?:[\s\S]*?\bfrom\s*)?["'](\.\.\/[^"']+)["']/gm),
-    ].map((m) => m[1].replace(/^\.\.\//, ""));
+      ...entrypointSrc.matchAll(/^\s*import\b(?:[\s\S]*?\bfrom\s*)?["'](\.{1,2}\/[^"']+)["']/gm),
+    ].map((m) => path.normalize(path.join("src", m[1])));
 
     for (const file of localImports) {
       const covered =
